requirements.txt
accelerate==0.31.0
data_processing==0.1.4
datasets==2.18.0
huggingface-hub==0.27.1
matplotlib==3.8.4
numpy==1.26.4
pandas==2.2.2
PyYAML==6.0.1
PyYAML==6.0.1
scipy==1.13.1
torch==2.2.2
tqdm==4.66.2
transformers==4.40.0
sentencepiece==0.2.0
protobuf==5.27.1
zstandard==0.22.0
ipdb==0.13.13
